FUNDALZHEIMER_CR first commit
FUNDALZHEIMER_CR

This is an electronic medical records system developed for the Costa Rican non-profit foundation FUNDALZHEIMER COSTA RICA, which treats people with Alzheimer's and other dementias.
The system allows the registration of patients with Alzheimer's and other dementias, as well as their caregivers (generally family, friends, relatives). It also has a reporting module where the information of a patient or caregiver registered in the SQL Server database is extracted and recorded in a "PDF" type file.
The system has user authentication of the Login Role type. This allows the user to have access only to the patient and caregiver record, while the administrator has access to the patient and caregiver record as well as to generate reports in "PDF".
System developed in Microsoft's C# high-level programming language.